+++ src/Form/Builder/TransitionFormBuilder.php
@@ -15,6 +15,8 @@ declare(strict_types=1);
 
 namespace Netzmacht\ContaoWorkflowBundle\Form\Builder;
 
+use Netzmacht\Workflow\Flow\Context;
+use Netzmacht\Workflow\Flow\Item;
 use Netzmacht\Workflow\Flow\Transition;
 use Symfony\Component\Form\FormBuilderInterface as FormBuilder;
 
@@ -36,9 +38,11 @@ interface TransitionFormBuilder
      * Build the transition form.
      *
      * @param Transition  $transition  The workflow transition.
+     * @param Item        $item        The current transition item.
+     * @param Context     $context     The current transition context.
      * @param FormBuilder $formBuilder The form builder.
      *
      * @return void
      */
-    public function buildForm(Transition $transition, FormBuilder $formBuilder): void;
+    public function buildForm(Transition $transition, Item $item, Context $context, FormBuilder $formBuilder): void;
 }