+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [Unreleased]
+
+### Breaking
+
+ - Changed transition form type. It requires the transition handler instead of the transition now
+ - Change interface of TransitionFormBuilder
+ - Only pass one transition form builder to the transition form type
+
 ## [2.0.0-rc3] - 2019-09-02
 
 ### Fixed